requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
Pillow
pandas
onnx
onnxruntime
tqdm